Use most narrow scope obj for Glide & fix title
Glide#with accepts Context, Activity and Fragment
objects and will follow the lifecycle of the
passed object to determine when to load images and
when to stop loading. By using the most narrow
scope, e.g. Fragment, then Activity, then Context
when available, we can help Glide make the most
informed decisions about when images should be
loaded.

Also fix title to not have underscores in Cards

Bug: 25864487
Change-Id: Ib6cafb2c791b9baa1b969775c71af84f7c244081