New database model - huge refactor
The following change refactors the database model to use
a single ContentProvider (VideoProvider) that is allows
access to all the app data through a Loader. As a result
there is no static VideoProvider and all data is accessed
from the database as needed.

All data is loaded into the database via FetchVideoService
which runs once on initial opening of the app.

The Movie class has been renamed to Video and made to be
immutable. The PlaybackOverlayActivity could not continue
to operate as-is using the new model, so everything was
migrated to the PlaybackOverlayFragment.

Bug: 25864165
Bug: 25864700
Bug: 25864506
Bug: 25864723
Bug: 25864802
Bug: 25864719
Bug: 25864237
Bug: 25864527

Change-Id: Ia3f1f6d1f2cdea481ac41b03934c94f89e8a652a